backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic-settings==2.10.1
boto3==1.40.23
celery==5.5.3
redis==6.4.0
requests==2.32.5
python-slugify==8.0.4
openai==2.11.0
